ci: Force colour output in detector run (#95)
# Pull Request

## Description

This change adds colour output to the log in the GitHub Actions workflow for deploying to GitHub Pages. Specifically, it sets the `FORCE_COLOR` environment variable to "true" for the "Run Detector" step, ensuring that the log output includes colour formatting. This enhancement improves the readability and visual clarity of the workflow logs.

fixes #93