ci: Add custom size labels to PR workflow (#139)
# Pull Request

## Description

This change enhances the pull request labelling process in the GitHub Actions workflow. It adds a name to the labeller step for improved readability and introduces size labelling for pull requests.

The size labelling action is configured with custom thresholds:
- XS: 0-19 lines
- S: 20-49 lines
- M: 50-199 lines
- L: 200-499 lines
- XL: 500-999 lines
- XXL: 1000+ lines

These updates will provide more detailed and automated categorisation of pull requests, helping to streamline the review process.

fixes #140