chore: Temporarily only allow scraping of JackPlowman/tech-detective (#153)
# Pull Request

## Description

This change modifies the `generate_tech_report` function in the `app.py`
file to focus on testing the scraper functionality. The previous
implementation, which scraped technologies for all repositories, has
been temporarily commented out.

The new code selects a specific repository named
"JackPlowman/tech-detective" from the list of repositories and applies
the `scrape_technologies` function to this single repository. This
targeted approach allows for more focused testing and debugging of the
scraper functionality.

The original loop that processed all repositories has been preserved as
a comment for future reimplementation, ensuring that the full
functionality can be easily restored once testing is complete.