chore: Add Debug Detector configuration to VS Code launch settings (#154)
# Pull Request

## Description

This change introduces a new debug configuration for the Detector module in the VS Code launch settings. The "Debug Detector" configuration allows for easier debugging of the application module within the detector directory.

Additionally, a minor adjustment has been made to the `scrape_technologies` function in the GitHub interactions utility. The redundant `return []` statement at the end of the function has been removed, as it was unnecessary due to the existing return statement.

These changes aim to improve the development experience by providing a dedicated debugging setup for the Detector module and streamlining the code in the GitHub interactions utility.

fixes #156